HW02 p5 (spring context)

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,5 +1,8 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.MealRepository;
@@ -10,11 +13,11 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
+@Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
 
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>(); // еда всех пользователей
-    //private Map<Integer, Meal> repositoryForAuthorizedUser = new ConcurrentHashMap<>(); // еда авторизованного пользователя
-
+    private static final Logger log = LoggerFactory.getLogger(InMemoryMealRepositoryImpl.class);
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
@@ -35,17 +38,29 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     }
 
     @Override
-    public void delete(int id, int userId) {
-        repository.remove(id);
+    public boolean delete(int id, int userId) {
+        log.info("delete {}", id);
+        if (repository.containsKey(id))
+        {
+            repository.remove(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
     public Meal get(int id, int userId) {
-        return repository.get(id);
+        log.info("get {}", id);
+        if (repository.containsKey(id))
+        {
+            return repository.get(id);
+        }
+        return null;
     }
 
     @Override
     public Collection<Meal> getAll(int userId) {
+        log.info("getAll");
         return repository.values();
     }
 }
